+++ api/Data/InstantRunoffContext.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
-using elections.Data;
+using shared;
 public partial class InstantRunoffContext : DbContext
 {
     public InstantRunoffContext()
