+++ shared/OfficeInElection.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace elections.Data
+namespace shared
 {
     public partial class OfficeInElection
     {
