+++ web/Pages/Databaseinit.cshtml
@@ -1,0 +1,6 @@
+﻿@page
+@model web.Pages.Index1Model
+@{
+}
+
+<h1>Hi</h1>
