+++ web/Pages/Databaseinit.cshtml.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace web.Pages
+{
+    public class Index1Model : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
